services/gateway/requirements.txt
fastapi==0.110.3
httpx==0.27.2
pydantic==2.9.2
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.8
slowapi==0.1.9
uvicorn==0.30.6
